refactor: move utils and types to a single file
